+++ src/main/java/com/minecolonies/coremod/compatibility/jei/ModRecipeTypes.java
@@ -1,6 +1,7 @@
 package com.minecolonies.coremod.compatibility.jei;
 
 import com.minecolonies.api.crafting.CompostRecipe;
+import com.minecolonies.coremod.colony.crafting.ToolUsage;
 import mezz.jei.api.recipe.RecipeType;
 
 import static com.minecolonies.api.util.constant.Constants.MOD_ID;
@@ -13,8 +14,8 @@ public class ModRecipeTypes
     public static final RecipeType<FishermanRecipeCategory.FishingRecipe> FISHING =
             RecipeType.create(MOD_ID, "fishing", FishermanRecipeCategory.FishingRecipe.class);
 
-    public static final RecipeType<ToolRecipeCategory.ToolUsage> TOOLS =
-            RecipeType.create(MOD_ID, "tools", ToolRecipeCategory.ToolUsage.class);
+    public static final RecipeType<ToolUsage> TOOLS =
+            RecipeType.create(MOD_ID, "tools", ToolUsage.class);
 
     private ModRecipeTypes()
     {
