+++ src/main/java/com/minecolonies/coremod/compatibility/jei/ToolRecipeCategory.java
@@ -1,10 +1,9 @@
 package com.minecolonies.coremod.compatibility.jei;
 
-import com.minecolonies.api.colony.IColonyManager;
 import com.minecolonies.api.items.ModItems;
-import com.minecolonies.api.util.ItemStackUtils;
-import com.minecolonies.api.util.constant.ToolType;
 import com.minecolonies.api.util.constant.TranslationConstants;
+import com.minecolonies.coremod.colony.crafting.ToolUsage;
+import com.minecolonies.coremod.colony.crafting.ToolsAnalyzer;
 import com.mojang.blaze3d.vertex.PoseStack;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -20,18 +19,16 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.FormattedText;
 import net.minecraft.network.chat.Style;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.Enchantments;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.*;
+import java.util.List;
 
 import static com.minecolonies.api.util.constant.Constants.MAX_BUILDING_LEVEL;
 
 /**
  * JEI recipe category showing supported tool and armor levels.
  */
-public class ToolRecipeCategory implements IRecipeCategory<ToolRecipeCategory.ToolUsage>
+public class ToolRecipeCategory implements IRecipeCategory<ToolUsage>
 {
     private static final int WIDTH = 180;
     private static final int HEIGHT = 44;
@@ -57,77 +54,7 @@ public class ToolRecipeCategory implements IRecipeCategory<ToolRecipeCategory.To
     @NotNull
     public static List<ToolUsage> findRecipes()
     {
-        final Map<ToolType, ToolUsage> toolItems = new HashMap<>();
-
-        for (final ItemStack stack : IColonyManager.getInstance().getCompatibilityManager().getListOfAllItems())
-        {
-            for (final ToolType tool : ToolType.values())
-            {
-                if (tool == ToolType.NONE || !ItemStackUtils.isTool(stack, tool)) { continue; }
-
-                tryAddingToolWithLevel(toolItems, tool, stack);
-
-                if (stack.isEnchantable())
-                {
-                    for (int enchantLevel = 1; enchantLevel < 4; ++enchantLevel)
-                    {
-                        tryAddingEnchantedTool(toolItems, tool, stack, enchantLevel);
-                    }
-                }
-            }
-        }
-
-        return toolItems.values().stream().sorted(Comparator.comparing(ToolUsage::tool)).toList();
-    }
-
-    private static void tryAddingEnchantedTool(@NotNull final Map<ToolType, ToolUsage> toolItems,
-                                               @NotNull final ToolType tool,
-                                               @NotNull final ItemStack stack,
-                                               final int enchantLevel)
-    {
-        final ItemStack enchantedStack = stack.copy();
-
-        // this list should theoretically end up applying a total of two enchants to each tool type
-        tryEnchantStack(enchantedStack, Enchantments.UNBREAKING, enchantLevel);
-        tryEnchantStack(enchantedStack, Enchantments.MOB_LOOTING, enchantLevel);
-        tryEnchantStack(enchantedStack, Enchantments.FLAMING_ARROWS, enchantLevel);
-        tryEnchantStack(enchantedStack, Enchantments.BLOCK_FORTUNE, enchantLevel);
-        tryEnchantStack(enchantedStack, Enchantments.ALL_DAMAGE_PROTECTION, enchantLevel);
-        tryEnchantStack(enchantedStack, Enchantments.FISHING_SPEED, enchantLevel);
-
-        tryAddingToolWithLevel(toolItems, tool, enchantedStack);
-    }
-
-    private static void tryEnchantStack(@NotNull final ItemStack stack,
-                                        @NotNull final Enchantment enchantment,
-                                        final int enchantLevel)
-    {
-        if (enchantment.canEnchant(stack) && enchantLevel >= enchantment.getMinLevel() && enchantLevel <= enchantment.getMaxLevel())
-        {
-            stack.enchant(enchantment, enchantLevel);
-        }
-    }
-
-    private static void tryAddingToolWithLevel(@NotNull final Map<ToolType, ToolUsage> toolItems,
-                                               @NotNull final ToolType tool,
-                                               @NotNull final ItemStack stack)
-    {
-        int level = ItemStackUtils.getMiningLevel(stack, tool);
-        if (level < 0) {
-            return;
-        }
-        level = Math.min(MAX_BUILDING_LEVEL, level + ItemStackUtils.getMaxEnchantmentLevel(stack));
-
-        final ToolUsage usage = toolItems.computeIfAbsent(tool, ToolUsage::create);
-
-        if (stack.isEnchanted())
-        {
-            usage.enchantedToolLevels.get(level).add(stack);
-        }
-        else
-        {
-            usage.toolLevels.get(level).add(stack);
-        }
+        return ToolsAnalyzer.findTools();
     }
 
     @NotNull
@@ -208,29 +135,5 @@ public class ToolRecipeCategory implements IRecipeCategory<ToolRecipeCategory.To
             x += 18;
         }
         stack.popPose();
-    }
-
-    /**
-     * Describes a tool and its level compatibility.
-     * @param tool                  the tool type
-     * @param toolLevels            basic items accepted at each building level.
-     * @param enchantedToolLevels   (some) enchanted items accepted at each building level.
-     */
-    public record ToolUsage(@NotNull ToolType tool,
-                            @NotNull List<List<ItemStack>> toolLevels,
-                            @NotNull List<List<ItemStack>> enchantedToolLevels)
-    {
-        @NotNull
-        public static ToolUsage create(@NotNull final ToolType tool)
-        {
-            final List<List<ItemStack>> basicLevels = new ArrayList<>();
-            final List<List<ItemStack>> enchantedLevels = new ArrayList<>();
-            for (int i = 0; i <= MAX_BUILDING_LEVEL; ++i)
-            {
-                basicLevels.add(new ArrayList<>());
-                enchantedLevels.add(new ArrayList<>());
-            }
-            return new ToolUsage(tool, basicLevels, enchantedLevels);
-        }
     }
 }
